fix(RacingCarGameMain) : 메인 메소드를 컨트롤러로 사용
