--- src/main/java/Application.java
@@ -1,8 +1,0 @@
-import racingCar.Game;
-
-public class Application {
-    public static void main(String[] args) {
-        Game game = new Game();
-        game.run();
-    }
-}
